loads of changes related to font handling, exception handling, and minor work on image handling

+++ ptext/object/canvas/geometry/line_segment.py
@@ -1,23 +1,24 @@
+from decimal import Decimal
 from math import sqrt
 
 
 class LineSegment:
-    def __init__(self, x0: float, y0: float, x1: float, y1: float):
+    def __init__(self, x0: Decimal, y0: Decimal, x1: Decimal, y1: Decimal):
         self.x0 = x0
         self.y0 = y0
         self.x1 = x1
         self.y1 = y1
 
-    def length(self) -> float:
-        return sqrt((self.x0 - self.x1) ** 2 + (self.y0 - self.y1) ** 2)
+    def length(self) -> Decimal:
+        return Decimal(sqrt((self.x0 - self.x1) ** 2 + (self.y0 - self.y1) ** 2))
 
-    def get_start(self) -> (float, float):
+    def get_start(self) -> (Decimal, Decimal):
         return (self.x0, self.y0)
 
-    def get_end(self) -> (float, float):
+    def get_end(self) -> (Decimal, Decimal):
         return (self.x1, self.y1)
 
-    def get_bounding_box(self) -> (float, float, float, float):
+    def get_bounding_box(self) -> (Decimal, Decimal, Decimal, Decimal):
         """
         return the bounding box of this rectangle in the format
         (lower_left_x, lower_left_y, width, height)
